--- content/geophysical_surveys/MT/index.rst
@@ -1,9 +1,0 @@
-.. _MT_index:
-
-MagnetoTelluric Method
-======================
-
-.. toctree::
-    :maxdepth: 1
-
-    MT_N_layered_Earth
