+++ content/geophysical_surveys/index.rst
@@ -11,7 +11,6 @@ Geophysical Surveys
     fundamentals/index
     dcr/index
     looploopfdem/index
-    MT/index
 
 
 For each geophysical method, we will follow the framework:
