+++ content/maxwell3_fdem/index.rst
@@ -9,3 +9,4 @@ Contents:
     :maxdepth: 1
 
     theory_3loops
+    MT/index
